add download progress logic
